Add smart-default to create easier enum defaults
This allows us to set an Enum::default() more easily, so we can follow the original code's use of "default" enum entries, as in many of the enum types in the original asm.h.
Eventually this might be part of the language, but for now we use this custom crate.

Related discussion: https://internals.rust-lang.org/t/request-derive-enums-default/10576