Generate symbol list as a String
Instead of directly writing the symbol list to a file or stdout, we generate it as a String. Specific functions then write it to a file with fputs(), or to stdout with println!(). This fixes an issue where printing to stdout was coming out of order due to the mixed use of fputs()/fprintf() and println!().