Moved mnemonics to a separate "mnemonics" module and cleaned up operation function call
The new "mnemonics" module contains all basic mnemonics operations, as well as the old "processor" operations.

The _MNE struct takes a simple function as a vect, instead of an <Option> of an unsafe extern function.

Many of the main.rs functions are now proper fn (although still unsafe) and not extern anymore.

An "operation" module for macros was also added.

This is an attempt to organize things better; many files were moved around, or split into two files. The result is a bit messy, but is a step in the right direction.