Moved *Av to the global state
This is a kind of argument buffer, used when parsing lines. This also meant changing the way some parse-related functions behave (e.g. findext) so the code doesn't have to rely on pointer side effects to rename arguments anymore.

In the future this should probably be revisited so we don't rely on floating arguments like that.